mm:swap: do not check readahead flag with THP anon
Huang reported that the PG_readahead flag uses PF_NO_COMPOUND so that we
cannot use the flag for THP page.  So, we need to check first whether the
page is THP before using TestClearPageReadahead in lookup_swap_cache().

This patch fixes it.

Furthermore, swap_[cluster|vma]_readahead cannot mark PG_readahead for
newly allocated page because the allocated page is always a normal page,
not THP at this moment.  So let's clean it up, too.

Link: http://lkml.kernel.org/r/874lm83zho.fsf@yhuang-dev.intel.com
Link: http://lkml.kernel.org/r/20180227232611.169883-1-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>