selftests: cgroup: add test for memory.low corner cases
Add a more complicated test for memory.low hierarchical behavior.  It
creates the following hierarchy:

        A memory.low = 50M
        B memory.low = 50M
        C memory.low = 50M
        D memory.low = 50M memory.(swap.)max = 200M
        E memory.low = 50M
       / \
      F'  F memory.low = 50M
          G memory.low = 50M
          H memory.low = 50M
          I memory.low = 50M
          J memory.low = 50M
          K memory.low = 50M, memory.usage = 50M

First, it creates local memory pressure by charging pagecache to F' and
checks that K's memory.low actually works (usage ~= 50M).  The it sets C's
memory.low to 0 and repeats the test to check that K's memory.low is not
working anymore, despite that memory.low is disabled above the cgroup with
memory pressure (D).

Link: http://lkml.kernel.org/r/20180522133106.24306-1-guro@fb.com
Signed-off-by: Roman Gushchin <guro@fb.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Greg Thelen <gthelen@google.com>
Cc: Tejun Heo <tj@kernel.org>
Cc: Shuah Khan <shuah@kernel.org>
Signed-off-by: Andrew Morton <akpm@linuxfoundation.org>