oom, memcg: clarify root memcg oom accounting
David Rientjes has pointed out that the current way how the root memcg is
accounted for the cgroup aware OOM killer is undocumented.  Unlike regular
cgroups there is no accounting going on in the root memcg (mostly for
performance reasons).  Therefore we are suming up oom_badness of its
tasks.  This might result in an over accounting because of the
oom_score_adj setting.  Document this for now.

Link: http://lkml.kernel.org/r/20180130122011.GB21609@dhcp22.suse.cz
Signed-off-by: Michal Hocko <mhocko@suse.com>
Acked-by: Roman Gushchin <guro@fb.com>
Acked-by: Tejun Heo <tj@kernel.org>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Tetsuo Handa <penguin-kernel@i-love.sakura.ne.jp>
Signed-off-by: Andrew Morton <akpm@linuxfoundation.org>