mm, devm_memremap_pages: mark devm_memremap_pages() EXPORT_SYMBOL_GPL
Patch series "mm: rework hmm to use devm_memremap_pages".

For maintainability, as ZONE_DEVICE continues to attract new users, it is
useful to keep all users consolidated on devm_memremap_pages() as the
interface for create "device pages".

The devm_memremap_pages() implementation was recently reworked to make it
more generic for arbitrary users, like the proposed peer-to-peer PCI-E
enabling.  HMM pre-dated this rework and opted to duplicate
devm_memremap_pages() as hmm_devmem_pages_create().

Rework HMM to be a consumer of devm_memremap_pages() directly and fix up
the licensing on the exports given the deep dependencies on the mm.


This patch (of 5):

The devm_memremap_pages() facility is tightly integrated with the kernel's
memory hotplug functionality.  It injects an altmap argument deep into the
architecture specific vmemmap implementation to allow allocating from
specific reserved pages, and it has Linux specific assumptions about page
structure reference counting relative to get_user_pages() and
get_user_pages_fast().  It was an oversight that this was not marked
EXPORT_SYMBOL_GPL from the outset.

Link: http://lkml.kernel.org/r/152694211948.5484.5084180448502160279.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Cc: Michal Hocko <mhocko@suse.com>
Cc: "Jérôme Glisse" <jglisse@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>