mm: list_lru: add lock_irq member to __list_lru_init()
scan_shadow_nodes() is the only user of __list_lru_walk_one() which
disables interrupts before invoking it.  The reason is that nlru->lock is
nesting inside IRQ-safe i_pages lock.  Some functions unconditionally
acquire the lock with the _irq() suffix.

__list_lru_walk_one() can't acquire the lock unconditionally with _irq()
suffix because it might invoke a callback which unlocks the nlru->lock
and invokes a sleeping function without enabling interrupts.

Add an argument to __list_lru_init() which identifies whether the
nlru->lock needs to be acquired with disabling interrupts or without.

Link: http://lkml.kernel.org/r/20180622151221.28167-4-bigeasy@linutronix.de
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Kirill Tkhai <ktkhai@virtuozzo.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>