iomap: use non-raw io functions for io{read|write}XXbe
Patch series "Add io{read|write}64 to io-64-atomic headers", v18.

A cleanup series to push a number of instances of people defining their
own io{read|write}64 functions into common headers seing they don't exist
in non-64bit systems.  This series adds inline functions to the
io-64-nonatomic headers and then cleans up the drivers that defined their
own copies.

This cleanup was originally requested by Greg after he reviewed my
Switchtec NTB code.


This patch (of 7):

Fix an asymmetry in the io{read|write}XXbe functions in that the
big-endian variants make use of the raw io accessors while the
little-endian variants use the regular accessors.  Some architectures
implement barriers to order against both spinlocks and DMA accesses and
for these case, the big-endian variant of the API would not be protected.

Thus, change the mmio_XXXXbe macros to use the appropriate swab() function
wrapping the regular accessor.  This is similar to what was done for PIO.

When this code was originally written, barriers in the IO accessors were
not common and the accessors simply wrapped the raw functions in a
conversion to CPU endianness.  Since then, barriers have been added in
some architectures and are now missing in the big endian variant of the
API.

This also manages to silence a few sparse warnings that check for using
the correct endian types which the original code did not annotate
correctly.

Link: http://lkml.kernel.org/r/20180622194752.11221-2-logang@deltatee.com
Link: http://lkml.kernel.org/r/CAK8P3a25zQDxyaY3iVv+JmSSzs7F6ssGc+HdBkGs54ZfViX+Fg@mail.gmail.com
Signed-off-by: Logan Gunthorpe <logang@deltatee.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Kate Stewart <kstewart@linuxfoundation.org>
Cc: Philippe Ombredanne <pombredanne@nexb.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Alan Cox <gnomes@lxorguk.ukuu.org.uk>
Cc: Allen Hubbe <Allen.Hubbe@dell.com>
Cc: Andy Shevchenko <andy.shevchenko@gmail.com>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Dan Douglass <dan.douglass@nxp.com>
Cc: Dave Jiang <dave.jiang@intel.com>
Cc: "David S. Miller" <davem@davemloft.net>
Cc: Helge Deller <deller@gmx.de>
Cc: Herbert Xu <herbert@gondor.apana.org.au>
Cc: Horia Geantă <horia.geanta@nxp.com>
Cc: "James E.J. Bottomley" <jejb@parisc-linux.org>
Cc: Jon Mason <jdmason@kudzu.us>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Nicholas Piggin <npiggin@gmail.com>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>